+++ packages/custodyKeyring/src/index.ts
@@ -1,3 +1,5 @@
 export { CUSTODIAN_TYPES } from "./custodianTypes";
 export { CustodyKeyring } from "./CustodyKeyring";
 export { Migrator } from "./migrations/migrator";
+export { MmiConfigurationController } from "./mmiConfiguration";
+export { MMIConfiguration } from "./types/MMIConfiguration";
