+++ packages/custodyKeyring/src/interfaces/IConfiguration.ts
@@ -1,0 +1,13 @@
+import { IJsonRpcCustodian } from "./IJsonRpcCustodian";
+
+export interface IConfiguration {
+  portfolio: {
+    enabled: boolean;
+    url: string;
+    cookieSetUrls: string[];
+  };
+  features: {
+    websocketApi: boolean;
+  };
+  custodians: IJsonRpcCustodian[];
+}
