+++ packages/custodyKeyring/src/interfaces/ICustodyKeyringOptions.ts
@@ -1,5 +1,5 @@
 import { ICustodianAccount } from "@metamask-institutional/types";
-import { MmiConfigurationController } from "@metamask-institutional/mmi-configuration";
+import { MmiConfigurationController } from "../mmiConfiguration";
 
 export interface ICustodyKeyringOptions {
   accounts?: string[];
