+++ packages/custodyKeyring/src/interfaces/IMmiConfigurationControllerOptions.ts
@@ -1,0 +1,17 @@
+import { IJsonRpcCustodian } from "./IJsonRpcCustodian";
+
+export class IMmiConfigurationControllerOptions {
+  mmiConfigurationServiceUrl?: string;
+  initState?: {
+    mmiConfiguration: {
+      portfolio: {
+        enabled: boolean;
+        url: string;
+      };
+      features: {
+        websocketApi: boolean;
+      };
+      custodians: IJsonRpcCustodian[];
+    };
+  };
+}
