+++ packages/custodyKeyring/src/types/MMIConfiguration.ts
@@ -1,0 +1,14 @@
+import { IJsonRpcCustodian } from "../interfaces/IJsonRpcCustodian";
+
+export type MMIConfiguration = {
+  mmiConfiguration: {
+    portfolio: {
+      enabled: boolean;
+      url: string;
+    };
+    features: {
+      websocketApi: boolean;
+    };
+    custodians: IJsonRpcCustodian[];
+  };
+};
