+++ packages/mmiConfiguration/src/index.ts
@@ -1,3 +1,1 @@
-export { MmiConfigurationController } from "./mmiConfiguration";
 export { ConfigurationClient } from "./ConfigurationClient";
-export { MMIConfiguration } from "./types";
