+++ packages/sdk/package.json
@@ -37,7 +37,6 @@
     "jsonwebtoken": "^9.0.0"
   },
   "devDependencies": {
-    "@types/jest": "^29.2.6",
     "@types/mocha": "^10.0.1",
     "jest-fetch-mock": "^3.0.3",
     "ts-jest": "26.4.4"
